requirements.txt
dbt-core 
dbt-postgres
ipykernel
pandas
matplotlib
sqlalchemy
python-dotenv